% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{addAgent}
\alias{addAgent}
\title{add an agent to a population}
\arguments{
\item{population}{an external pointer to a population, for example,
one returned by \code{\link[=newPopulation]{newPopulation()}}}

\item{agent}{an external pointer to an agent, returned by \code{\link[=newAgent]{newAgent()}} or
\code{\link[=getAgent]{getAgent()}}}
}
\description{
add an agent to a population
}
\details{
if the agent is an R6 class, we should use \code{agent$get} to get
the external pointer. Similarly, if population is an R6 object, then we
should either use \code{population$addAgent()} or \code{population$get}.
}
