\name{ABM-package}
\alias{ABM-package}
\alias{ABM}
\docType{package}
\title{
\packageTitle{ABM}
}
\description{
\packageDescription{ABM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ABM}
\packageIndices{ABM}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{ABM}

Maintainer: \packageMaintainer{ABM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
