% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agent.R
\name{clearEvents}
\alias{clearEvents}
\title{Unschedule all event from an agent}
\usage{
clearEvents(agent)
}
\arguments{
\item{agent}{an external pointer returned by newAgent}
}
\value{
the agent itself
}
\description{
Unschedule all event from an agent
}
\details{
If agent is an R6 object, then we should either use agent$schedule,
or use schedule(agent$get, event)
}
