% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_network_functions.R
\name{comm_loc}
\alias{comm_loc}
\title{Community location}
\usage{
comm_loc(comm, nbhd_mat_list)
}
\arguments{
\item{comm}{community (neighborhood) of residence}

\item{nbhd_mat_list}{list of neighbors for all neighborhoods}
}
\value{
location of agent if they choose to be in the community
}
\description{
Function to randomly determine location if agent is not at work, home, or school. 10% chance of visiting neighboring community, 1% chance of visiting some random community.
}
