% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_V1.R, R/ABM_V2.R
\name{covid_abm}
\alias{covid_abm}
\title{COVID ABM}
\usage{
covid_abm(
  bta,
  E0,
  Ip0,
  Ia0,
  Im0,
  Imh0,
  Ih0,
  R0,
  D0,
  agents,
  nbhd_mat_list,
  t.tot,
  dt,
  day_of_week_fx,
  time_of_day_fx,
  SiP_fx,
  scl_fx,
  test_fx
)

covid_abm(
  bta,
  E0,
  Ip0,
  Ia0,
  Im0,
  Imh0,
  Ih0,
  R0,
  D0,
  agents,
  nbhd_mat_list,
  t.tot,
  dt,
  day_of_week_fx,
  time_of_day_fx,
  SiP_fx,
  scl_fx,
  test_fx
)
}
\arguments{
\item{bta}{beta parameter = transmission probability}

\item{E0}{initial number of exposed agents, to be allocated among working population}

\item{Ip0}{initial number of pre-symptomatic agents, to be allocated among working population}

\item{Ia0}{initial number of asymptomatic agents, to be allocated among working population}

\item{Im0}{initial number of mildly symptomatic agents, to be allocated among working population}

\item{Imh0}{initial number of mildly symptomatic-pre-hospitalized agents, to be allocated among working population}

\item{Ih0}{initial number of hospitalized agents, to be allocated among working population}

\item{R0}{initial number of recovered agents, to be allocated among working population}

\item{D0}{initial number of deceased agents, to be allocated among working population}

\item{agents}{data.table of synthetic agents, must contain columns `age`, `race`, `state`, `nextstate`, `tnext`, `t_infection`, `test_prob`, `tested`, `t_since_test`, `q_prob`, `sociality`, `residence`, `residence_type`, `res_inf`, `income`, `work`, `school`, `nbhd`, `office_id`, `class_id`, `essential`, `comm_bracket`, and `income_bracket`}

\item{nbhd_mat_list}{list derived from neighborhoods matrix with each neighborhoods neighbors. See `comm_loc` and `GetNbhd` functions}

\item{t.tot}{total time period in days to run simulation}

\item{dt}{time step of simulation, default is 4/24. Be sure to match day_of_week_fx and time_of_day_fx appropriately since total time steps will equal `t.tot/dt`}

\item{day_of_week_fx}{function returning the day of week of at time step `t`}

\item{time_of_day_fx}{function returning the time of day at time step `t`}

\item{SiP_fx}{function returning shelter-in-place status of population at `t`}

\item{scl_fx}{function returning school closure status at time step `t`}

\item{test_fx}{function returning the number of tests conducted at time step `t`}

\item{pop.ages}{vector corresponding to age of each individual}

\item{q.probs}{vector of individual probabilities of quarantining}

\item{n.probs}{vector of individual probabilities of being in random edge generation}

\item{inf.mat}{infection status matrix with column 1 filled with initial infection status, N rows, and t.tot/dt columns}

\item{transition.mat}{transition times matrix (time until moving into next state) with column 1 filled with initial times remaining in starting states, N rows, and t.tot/dt columns}

\item{net.mat}{contact network array with dimensions NxNxt.tot/dt such that each slice is a symmetric matrix representing contact between}
}
\value{
Matrix with columns corresponding to each infection state and rows filled with number people in each state at each time over which the simulation was run


}
\description{
Simulate COVID transmission across a network of N individuals

Simulate agents based model of COVID19 transmission from a population of synthetic agents
}
