\name{temperatures}
\alias{temperatures}
\docType{data}
\title{
Outside slurry temperatures and weather
}

\description{
Slurry temperatures according to the Danish National Inventory Report and from Vechi et al. 2023.
}

\usage{
data(temperatures)
}

\format{
The format is: \rda
\code{weather_dat_DK}
\code{weather_dat_DK}

}
}
\source{
Data from Sound clips were recorded in Vermont, USA in 2010.  Equipment was a Wildlife Acoustics SM1(TM) recorder recording in WAC0 format, converted to wave using the Wildlife Acoustics Wac2Wav (TM) converter.  Recording has a sample rate of 24kHz and is 16-bit mono.
}

\examples{
data(temperatures)

viewSpec(temperatures)
}

\keyword{datasets}