% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_Helpers.R
\name{get.degree}
\alias{get.degree}
\title{Network degree}
\usage{
get.degree(net.mat)
}
\arguments{
\item{net.met}{matrix representing contact network}
}
\value{
vector of individual degrees (number of contacts)
}
\description{
get distribution of individual degrees from input network matrix
}
