% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_infection_functions.R
\name{get_foi}
\alias{get_foi}
\title{Person's FOI in large areas}
\usage{
get_foi(I, N, bta)
}
\arguments{
\item{I}{Location of this person}

\item{N}{number f infectious people transmitting}

\item{bta}{transmission probability}
}
\value{
FOI for this person
}
\description{
Determine FOI from locations with infectious individuals in large areas: schools, workplaces, communities (as opposed to corresponding small areas: classrooms, offices, homes)
}
