% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_infection_functions.R
\name{get_foi_small_area}
\alias{get_foi_small_area}
\title{(DEPRECATED) Person's FOI in small areas}
\usage{
get_foi_small_area(
  location,
  n_transmitting,
  res_size,
  office_size,
  class_size,
  trans_prob
)
}
\arguments{
\item{location}{Location of this person}

\item{n_transmitting}{number f infectious people transmitting}

\item{res_size}{population of this person's residence}

\item{office_size}{population of this person's office}

\item{class_size}{population of this person's class}

\item{trans_prob}{transmission probability}
}
\value{
FOI for this person
}
\description{
Determine FOI from locations with infectious individuals in small areas: classrooms, offices, homes
}
