% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{getAgent}
\alias{getAgent}
\title{Get the agent at an index in the population}
\arguments{
\item{population}{an external pointer to a population, for example,
one returned by \code{\link[=newPopulation]{newPopulation()}}}

\item{i}{the index of the agent, starting from 1.}
}
\value{
the agent at index i in the population.
}
\description{
Get the agent at an index in the population
}
