% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{getSize}
\alias{getSize}
\title{Get the size of a population}
\arguments{
\item{population}{an external pointer to a population, for example,
one returned by \code{\link[=newPopulation]{newPopulation()}}}
}
\value{
the population size, an integer
}
\description{
Get the size of a population
}
