% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agent.R
\name{getState}
\alias{getState}
\title{Get the state of the agent}
\arguments{
\item{agent}{an external pointer returned by newAgent}
}
\value{
a list holding the state
}
\description{
Get the state of the agent
}
\details{
If agent is an R6 object, then we should either use agent$schedule,
or use schedule(agent$get, event)
}
