% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event.R
\name{getTime}
\alias{getTime}
\title{returns the event time}
\arguments{
\item{event}{an external pointer returned by the newEvent function.}
}
\value{
a numeric value

This function avoids the overhead of an R6 class, and is thus faster.
This is the recommended method to get event time in an event handler.
}
\description{
returns the event time
}
