% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transition.R
\name{getWaitingTime}
\alias{getWaitingTime}
\title{Generate a waiting time from an WaitingTime object}
\arguments{
\item{generator}{an external pointer to a WaitingTime object, e.g.,
one returned by newExpWaitingTime or newGammaWaitingTime}

\item{time}{the current simulation time, a numeric value}
}
\value{
a numeric value
}
\description{
Generate a waiting time from an WaitingTime object
}
