% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventLogger.R
\name{inc}
\alias{inc}
\title{Create an event logger that increments a simulation state variable}
\arguments{
\item{variable}{the name of a numeric state variable in the simulation}

\item{filter}{an optional function receiving the state of the event agent}
}
\value{
An event logger that can be passed in the \code{logging} argument of
\code{Simulation$addTransition()}.
}
\description{
Creates a C++ event logger that increments a named simulation state variable
after each successful transition.
}
