% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agent.R
\name{leave}
\alias{leave}
\title{leave the population that the agent is in}
\arguments{
\item{agent}{an external pointer returned by newAgent}
}
\description{
leave the population that the agent is in
}
\details{
If agent is an R6 object, then we should use either
agent$leave() or leave(agent$get)
}
