% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_network_functions.R
\name{location_small}
\alias{location_small}
\title{Small location}
\usage{
location_small(l_id, work_id, school_id, office_id, class_id)
}
\arguments{
\item{l_res}{residence location id}

\item{l_com}{community location id}

\item{l_scl}{work/school location id}

\item{p_res}{probability of being at residence}

\item{p_com}{probability of being in community}
}
\value{
vector of location ids
}
\description{
Function to allocate agent locations to smaller subunits( offices and classes) from larger areas (workplaces and schools)
}
