% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agent.R
\name{matchState}
\alias{matchState}
\title{Check if the state of an agent matches a given state}
\usage{
matchState(agent, rule)
}
\arguments{
\item{agent}{an external pointer returned by newAgent}

\item{rule}{a list holding the state to match against}
}
\value{
a logical value
}
\description{
Check if the state of an agent matches a given state
}
\details{
This function is equivalent to
stateMatch(getState(agent), rule)

The state matches the rule if and only if each domain (names of the
list) in rule has the same value as in state. The domains in domains of the
state not listed in rule are not matched
}
