% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agent.R
\name{newAgent}
\alias{newAgent}
\title{Create an agent with a given state}
\arguments{
\item{state}{a list giving the initial state of the agent, or NULL (an empty
list)}

\item{death_time}{the death time for the agent, an optional numeric value.}
}
\value{
an external pointer pointing to the agent
}
\description{
Create an agent with a given state
}
\details{
Setting death_time is equivalent to calling the \code{\link[=setDeathTime]{setDeathTime()}}
function.
}
