% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event.R
\name{newEvent}
\alias{newEvent}
\title{Creates a new event in R}
\arguments{
\item{time}{the time that this event will occur. A length-1
numeric vector.}

\item{handler}{an R function that handles the event when it occurs.}
}
\value{
an external pointer, which can then be passed to
functions such as schedule and unschedule.
}
\description{
Creates a new event in R
}
\details{
The R handler function should take exactly 3 arguments
\enumerate{
\item time: the current time in the simulation
\item sim: the simulation object, an external pointer
\item agent: the agent to whom this event is attached to.
}

The return value of the handler function is ignored.

This function avoids the overhead of an R6 class, and is thus faster.
This is the recommended method to create an event in an event handler.
}
