% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transition.R
\name{newGammaWaitingTime}
\alias{newGammaWaitingTime}
\title{Creates an gamma distributed waiting time}
\arguments{
\item{shape}{the shape parameter of the gamma distribution}

\item{scale}{the scale parameter of the gamma distribution, i.e., 1/rate}
}
\value{
an external pointer
}
\description{
Creates an gamma distributed waiting time
}
\details{
This function creates an C++ object of type ExpWaitingTime.
It can be passed to addTransition or Simulation$addTransition to
specify the waiting time for a transition. As a C++ object, it is faster
than using an R function to generate waiting times because there is
no need to call an R function from C++.
}
