% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{newPopulation}
\alias{newPopulation}
\title{Create a new population}
\arguments{
\item{n}{an integer specifying the population size.}
}
\description{
Create a new population
}
\details{
The population will be created with "n" individuals in it.
These individuals have an empty state upon created. Note that
individuals can be added later by the "add" method, the initial
population size is for convenience, not required
}
