% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_infection_functions.R
\name{next.state}
\alias{next.state}
\title{Next State}
\usage{
next.state(pre.status, age)
}
\arguments{
\item{pre.status}{Infection status that has expired}

\item{age}{age of individual for probability of moving to different states}
}
\value{
vector with two entries, the first is the next state, the second is the time to be spent in that state. Time should be converted to `numeric` for subsequent use
}
\description{
Take input infection state and age and return the next infection state and the time spent in that state
}
