% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_network_functions.R
\name{other_location}
\alias{other_location}
\title{Simulate location of individuals who are not workers or school-aged}
\usage{
other_location(
  inf.state,
  tested,
  SiP,
  time_day,
  age,
  sociality,
  res_type,
  comm_bracket,
  res_id,
  comm_id
)
}
\arguments{
\item{inf.state}{Infection state, one of S, E, Ip, Ia, Im, Imh, Ih, R}

\item{tested}{test status of individual 1/0}

\item{SiP}{shelter in place active? 1/0}

\item{time_day}{time of day (night or day)}

\item{age}{age of person}

\item{sociality}{agent sociality}

\item{res_type}{type of residence (H, C, P, or N)}

\item{comm_bracket}{income bracket of the community (census tract)}

\item{res_id}{id of this individual's residence}

\item{comm_id}{id of this individual's community}
}
\value{
location of this individual in the time step with options being the id corresponding to their home (H), work (W), school (S), or community (C)
}
\description{
Simulate location of school-aged children depending on agent characteristics, NPIs, and time of day
}
