% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_infection_functions.R
\name{p_symp}
\alias{p_symp}
\title{Probability of symptomatic infection}
\usage{
p_symp(age)
}
\arguments{
\item{age}{age of person}
}
\value{
numeric of probability of having symptoms.
}
\description{
Age-dependent probabilities of having symptoms. Used in `next.state`
}
