\name{parameters}
\alias{parameters}
\docType{data}
\title{
  Default and alternativ parameter sets for ABM
}

\description{
  Suggested parameter sets and versions of parameter sets used in ABM
}

\usage{

  data(grp_pars2.0), the grp_pars default parameter set (v2.0) 
  data(grp_pars1.0), An alternative grp_pars parameter set (v1.0) used in Dalby et al. 2023, DOI: 10.1021/acs.est.2c08891
  data(man_pars2.0), the man_pars default parameter set which is based on an average Danish pig diet and with substrates defiend as RFd, CP, Cfat and starch rather than VSd
  data(man_pars1.0), An alternative man_pars parameter set (v1.0) based on VSd as substrate input.
  data(mic_pars2.0), the mic_pars default parameter set (v2.0)
  data(wthr_pars2.0), the wthr_pars default parameter set (v2.0)
  data(chem_pars2.0), the chem_pars default parameter set (v2.0)
  data(arrh_pars2.0), the arrh_pars default parameter set (v2.0)
  
  
}

\format{
  The format is: \rda
  
}
}
\source{
  
}

\examples{

  viewSpec(man_pars2.0)
  viewSpec(grp_pars2.0)
}

\keyword{datasets}