% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_testing_functions.R
\name{quar_prob}
\alias{quar_prob}
\title{Quarantine probability}
\usage{
quar_prob(inf.status, income, age, race, res_type, hh.inf)
}
\arguments{
\item{inf.status}{Infection status}

\item{income}{income}

\item{age}{age}

\item{res_type}{type of residence}

\item{hh.inf}{hh infection status}

\item{test.status}{hh infection status}
}
\value{
sampling weight for whether or not tested
}
\description{
Estimate probability someone will stay home
}
