% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_infection_functions.R
\name{quarantine}
\alias{quarantine}
\title{ABM quarantine}
\usage{
quarantine(inf.vec, test.vec, q.prob)
}
\arguments{
\item{inf.vec}{vector of individual infection states}

\item{test.vec}{vector of individual testing status}

\item{q.prob}{vector of individual probabilities of being quarantined}
}
\value{
vector of NAs and Ts with Ts corresponding to individuals who tested positive
}
\description{
Quarantine individuals based on their test and infection status
}
