% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_network_functions.R
\name{quarantine_symptomatics}
\alias{quarantine_symptomatics}
\title{Quarantine symptomatics}
\usage{
quarantine_symptomatics(inf.stat, base.net, q.prob)
}
\arguments{
\item{inf.stat}{Infection status vector}

\item{base.net}{Relationship matrix}

\item{q.prob}{Vector of length N corresponding to individuals' probability of quarantining given symptoms}
}
\value{
Updated contact matrix with quarantining of symptomatics implemented
}
\description{
Restrict contact of people with symptoms to their household with some compliance probability influenced by whether they're able to work from home
}
