% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_network_functions.R
\name{sac_location}
\alias{sac_location}
\title{Simulate school-aged children location}
\usage{
sac_location(
  inf.state,
  tested,
  scl,
  SiP,
  time_day,
  day_week,
  comm_bracket,
  age,
  sociality,
  res_id,
  scl_id,
  comm_id
)
}
\arguments{
\item{inf.state}{Infection state, one of S, E, Ip, Ia, Im, Imh, Ih, R}

\item{tested}{test status of individual 1/0}

\item{scl}{schools closed? 1/0}

\item{SiP}{shelter in place active? 1/0}

\item{time_day}{time of day (night or day)}

\item{day_week}{day of the week (U, M, T, W, R, F, or S)}

\item{comm_bracket}{income bracket of the community (census tract)}

\item{age}{age of person}

\item{sociality}{agent sociality}

\item{res_id}{id of this individual's residence}

\item{scl_id}{id of this individual's school}

\item{comm_id}{id of this individual's community}
}
\value{
location of this individual in the time step with options being the id corresponding to their home (H), school (S), or community (C)
}
\description{
Simulate location of school-aged children depending on agent characteristics, NPIs, and time of day
}
