% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agent.R
\name{schedule}
\alias{schedule}
\title{Schedule (attach) an event to an agent}
\arguments{
\item{agent}{an external pointer returned by newAgent}

\item{event}{an external pointer returned by newEvent}
}
\description{
Schedule (attach) an event to an agent
}
\details{
If agent is an R6 object, then we should use either
agent$schedule(event) or schedule(agent$get, event)

Similarly, if event is an R6 object, then we should use
schedule(agent, event$get)
}
