% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agent.R
\name{setDeathTime}
\alias{setDeathTime}
\title{set the time of death for an agent}
\arguments{
\item{agent}{an external pointer returned by \code{\link[=newAgent]{newAgent()}} or \code{\link[=getAgent]{getAgent()}}}

\item{time}{the time of death, a numeric value}
}
\description{
set the time of death for an agent
}
\details{
If agent is an R6 object, then we should use either
agent$leave() or leave(agent$get)

At the time of death, the agent is removed from the simulation. Calling it
multiple times causes the agent to die at the earliest time.
}
