% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agent.R
\name{setState}
\alias{setState}
\title{Merge values into the state of the agent}
\arguments{
\item{agent}{an external pointer returned by newAgent}

\item{state}{an R list giving the state values to update. Existing named
domains are replaced and new named domains are added; other domains remain
unchanged.}
}
\description{
Merge values into the state of the agent
}
\details{
In this framework, a state is a list, each named
component is called a domain. This function replaces values for
matching named domains and adds new named domains, while leaving
other components unchanged.

If agent is an R6 object, then we should either use agent$schedule,
or use schedule(agent$get, event)
}
