% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{setStates}
\alias{setStates}
\title{Set the state for each agent in a population}
\arguments{
\item{population}{an external pointer to a population, for example,
one returned by \code{\link[=newPopulation]{newPopulation()}}}

\item{states}{either a list holding the states (one for each agent), or a
function}
}
\description{
Set the state for each agent in a population
}
\details{
If \code{states} is a function then it takes a single argument
\code{i}, specifying the index of the agent (starting from 1), and returns
a state.
}
