% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM-package.R
\name{State}
\alias{State}
\title{The state of an agent}
\description{
In this framework, a state is a list, each named component is called a
domain. The value of a domain can be any R value. The list can be at most
one unnamed value, which corresponds to a domain with no name. This is
useful if there is only one domain.
}
\details{
A state can be matched to an R list (called a rule in this case).
The state matches the rule if and only if each domain (names of the
list) in rule has the same value as in state. The domains in domains of the
state not listed in rule are not matched. In addition, to match to a rule,
the domain values must be either a number or a character. This is useful
for identifying state changes. See \code{inc()}, \code{dec()}, and the
\link{Simulation} class' \code{addTransition} method for more details.
}
