% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_Helpers.R
\name{sum_inf_mat}
\alias{sum_inf_mat}
\title{Summarize Infection}
\usage{
sum_inf_mat(inf.mat)
}
\arguments{
\item{inf.mat}{infections status matrix from `ABM` run}
}
\value{
matrix with columns corresponding to each infection state and rows filled with number people in each state at each time
}
\description{
Function to summarize infection matric through time to give number of individuals in each state at each time
}
