% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_infection_functions.R
\name{t_asymp}
\alias{t_asymp}
\title{Asymptomatic period}
\usage{
t_asymp(shape.ia = 6, scale.ia = 1)
}
\arguments{
\item{shape.ia}{Shape of gamma distribution}

\item{scale.ia}{Scale of gamma distribution}
}
\value{
numeric of time spent in presymptomatic period (Ip)
}
\description{
Draw from distribution of proposed asymptomatic period. Used in `next.state`
}
