% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_infection_functions.R
\name{t_sevsymp}
\alias{t_sevsymp}
\title{time spent severely symptomatic (in hospital)}
\usage{
t_sevsymp(shape.ih = 3, scale.ih = 0.25)
}
\arguments{
\item{shape.ih}{shape of gamma distn}

\item{scale.ih}{scale of gamma distn}
}
\value{
numeric of time spent hospitalized (Ih)
}
\description{
Draw from distribution of proposed time spent with severe symptoms before dying or being discharged. Used in `next.state`
}
