% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_infection_functions.R
\name{test_folks}
\alias{test_folks}
\title{ABM Test}
\usage{
test_folks(inf.states, test.prior, test.probs, n.tests)
}
\arguments{
\item{inf.states}{vector of individual infection states}

\item{test.prior}{vector of prior testing statuses}

\item{test.probs}{vector of sampling weights corresponding to S, E, Ip, Ia, Im, Imh, Ih, and R classes, in that order}

\item{n.tests}{numeric of number of tests conducted}
}
\value{
vector of NAs and Ts with Ts corresponding to individuals who tested positive
}
\description{
Test individuals with different sampling weights. TODO: Implement testing uncertainty, esp. with regards to time since being infected
}
