% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_testing_functions.R
\name{test_sens}
\alias{test_sens}
\title{False Negative Rate}
\usage{
test_sens(inf.status, t.since.infection)
}
\arguments{
\item{inf.status}{Infection status}

\item{t.since.infection}{time since being infected}
}
\value{
binary of whether test is positive 1 or negative 0
}
\description{
Estimate probability of testing positive given true positive. Depends on function `pcr_sens_fun` which takes days since infection as input and returns probability of positive test given true positive.
}
