% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agent.R
\name{unschedule}
\alias{unschedule}
\title{Unschedule (detach) an event from an agent}
\arguments{
\item{agent}{an external pointer returned by newAgent}

\item{event}{an external pointer returned by newEvent}
}
\value{
the agent itself
}
\description{
Unschedule (detach) an event from an agent
}
\details{
If agent is an R6 object, then we should either use agent$schedule,
or use schedule(agent$get, event)

Similarly, if event is an R6 object, then we should use
schedule(agent, event$get)
}
