% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM-package.R
\docType{package}
\name{ABM-package}
\alias{ABM}
\alias{ABM-package}
\title{Agent Based Model Simulation Framework}
\description{
This package provides a framework to simulate agent based models that are
based on states and events.
}
\details{
\subsection{Agent}{

The concept of this framework is agent, which is an object of the \link{Agent}
class. An agent maintains its own state, which is a named R list storing any
R values in it. (see \link{State} ). The main task of an agent is to manage events
(see \link{Event}), and handle them in chronological order.
}

\subsection{Population}{

An object of the \link{Population} class manages agents and their contacts. The
contacts of agents are managed by Contact objects. The main functionality for
a contact object is to provide contacts of a given individuals at a given
time. For example, \code{\link[=newRandomMixing]{newRandomMixing()}} returns such an object that finds a
random agent in the population as a contact. the effect of contacts on the
states of agents are defined using a state transition rule. Please see
\code{addTransition} method of \link{Simulation} for more details.
}

\subsection{Simulation}{

The \link{Simulation} class inherits the \link{Population} class. So a simulation
manages agents and their contacts. Thus, the class also inherits the \link{Agent}
class. So a simulation can have its own state, and events attached
(scheduled) to it. In addition, it also manages all the transitions, using
its \code{addTransition} method. At last, it maintains loggers, which record
(or count) the state changes, and report their values at specified times.

During a simulation the earliest event in the simulation is picked out,
unscheduled (detached), and handled, which potentially causes the state
change of the agent (or another agent in the simulation). The transition can
update simulation state variables with event loggers such as \code{inc()} and
\code{dec()}, while report loggers such as
\code{\link[=newStateLogger]{newStateLogger()}} read values at requested times.
}

\subsection{Usage}{

To use this framework, we start by creating a simulation
object, populate the simulation with agents (either using the argument in
the constructor, or use its \code{addAgent} method), and
initialize the agents with their initial states using its \code{setState} method.

We then attach (\code{\link[=schedule]{schedule()}}) events to agents (possibly to the populations or
the simulation object too), so that these events change the agents' state.
For models which agents' states are defined by discrete states, such as the
SIR epidemic model, the events are managed by the framework through state
transitions, using rules defined by the \code{addTransition} method of
the \link{Simulation} class.

At last, we add report loggers to the simulation using
the \link{Simulation} class' \code{addLogger} method` and
\code{\link[=newStateLogger]{newStateLogger()}}. Event loggers are attached
to transitions through their \code{logging} argument. At last, run the simulation using
its \code{run} method, which returns the observations of the loggers
at the requested time points as a data.frame object.

For more information and examples, please see the \href{https://github.com/junlingm/ABM/wiki/}{Wiki} pages on Github.
}
}
\examples{
# simulate an SIR model using the Gillespie method
# the population size
N = 10000
# the initial number of infectious agents
I0 = 10
# the transmission rate
beta = 0.4
# the recovery rate
gamma = 0.2
# an waiting time egenerator that handles 0 rate properly
wait.exp = function(rate) {
  if (rate == 0) Inf else rexp(1, rate)
}
# this is a function that rescheduled all the events. When the
# state changed, the old events are invalid because they are
# calculated from the old state. This is possible because the
# waiting times are exponentially distributed
reschedule = function(time, agent, state) {
  clearEvents(agent)
  t.inf = time + wait.exp(beta*state$I*state$S/N)
  schedule(agent, newEvent(t.inf, handler.infect))
  t.rec = time + wait.exp(gamma*state$I)
  schedule(agent, newEvent(t.rec, handler.recover))
}
# The infection event handler
# an event handler take 3 arguments
#   time is the current simulation time
#   sim is an external pointer to the Simulation object.
#   agent is the agent that the event is scheduled to
handler.infect = function(time, sim, agent) {
  x = getState(agent)
  x$S = x$S - 1
  x$I = x$I + 1
  setState(agent, x)
  reschedule(time, agent, x)
}
# The recovery event handler
handler.recover = function(time, sim, agent) {
  x = getState(agent)
  x$R = x$R + 1
  x$I = x$I - 1
  setState(agent, x)
  reschedule(time, agent, x)
}
# create a new simulation with no agent in it.
# note that the simulation object itself is an agent
sim = Simulation$new()
# the initial state
x = list(S=N-I0, I=I0, R=0)
sim$state = x
# schedule an infection event and a recovery event
reschedule(0, sim$get, sim$state)
# add state loggers that saves the S, I, and R states
sim$addLogger("S")
sim$addLogger("I")
sim$addLogger("R")
# now the simulation is setup, and is ready to run
result = sim$run(0:100)
# the result is a data.frame object
print(result)

# simulate a standard agent-based SIR model
N = 100
I0 = 1
beta = 0.4
gamma = 0.2
sim = Simulation$new(
  N,
  function(i) list(stage = if (i <= I0) "I" else "S")
)
sim$state = list(S = N - I0, I = I0, R = 0)
S = list(stage = "S")
I = list(stage = "I")
R = list(stage = "R")
physical = "physical"
m = newRandomMixing(type = physical)
sim$addContact(m)
sim$addTransition(
  S + I -> I + I ~ physical,
  beta,
  logging = list(dec("S"), inc("I"))
)
sim$addTransition(
  I -> R,
  gamma,
  logging = list(dec("I"), inc("R"))
)
sim$addLogger("S")
sim$addLogger("I")
sim$addLogger("R")
# run the simulation, and get a data.frame object
result = sim$run(0:100)
print(result)

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/junlingm/ABM}
  \item Report bugs at \url{https://github.com/junlingm/ABM/issues}
}

}
\author{
\strong{Maintainer}: Junling Ma \email{junlingm@uvic.ca}

}
