% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agent.R
\name{clearEvents}
\alias{clearEvents}
\title{Unschedule all event from an agent}
\arguments{
\item{agent}{an external pointer returned by newAgent}
}
\description{
Unschedule all event from an agent
}
\details{
If agent is an R6 object, then we should use either
agent$clearEvents() or clearEvents(agent$get)
}
