% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_network_functions.R
\name{comm_loc}
\alias{comm_loc}
\title{Community location}
\usage{
comm_loc(comm)
}
\arguments{
\item{comm}{community (neighborhood) of residence}
}
\description{
Function to randomly determine location if agent is not at work, home, or school. 10% chance of visiting neighboring community, 1% chance of visiting some random community. Requires `sf.wmat` in environment in order to determine neighboring and random communities to potential visit
}
