% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_infection_functions.R
\name{get_foi}
\alias{get_foi}
\title{Person's FOI}
\usage{
get_foi(
  location,
  n_transmitting,
  res_size,
  work_size,
  school_size,
  comm_size,
  trans_rate
)
}
\arguments{
\item{location}{Location of this person}

\item{n_transmitting}{number f infectious people transmitting}

\item{res_size}{population of this person's residence}

\item{work_size}{population of this person's workplace}

\item{school_size}{population of this person's school}

\item{comm_size}{population of this person's community}
}
\value{
FOI for this person
}
\description{
Determine FOI from location
}
