% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contact.R
\name{newRandomMixing}
\alias{newRandomMixing}
\title{Creates a RandomMixing object}
\usage{
newRandomMixing()
}
\value{
an external pointer.
}
\description{
Creates a RandomMixing object
}
\details{
The population must be an external pointer, not an R6 object
To use an R6 object, we should use its pointer representation from its
$get method.
}
\examples{
# creates a simulation with 100 agents
sim = Simulation$new(100)
# add a random mixing contact pattern for these agents.
sim$addContact(newRandomMixing())
}
