% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contact.R
\name{newRandomMixing}
\alias{newRandomMixing}
\title{Creates a RandomMixing object}
\value{
an external pointer.
}
\description{
Creates a RandomMixing object
}
\examples{
# creates a simulation with 100 agents
sim = Simulation$new(100)
# add a random mixing contact pattern for these agents.
sim$addContact(newRandomMixing())

}
