% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agent.R
\name{setState}
\alias{setState}
\title{Set the state of the agent}
\arguments{
\item{agent}{an external pointer returned by newAgent}

\item{state}{an R list giving the components of the state to be
undated.}
}
\description{
Set the state of the agent
}
\details{
In this framework, a state is a list, each named
component is called a domain. This function only updates the
values of the domain given in the "value" list, while leave the
other components not in the "value" list unchanged.

If agent is an R6 object, then we should either use agent$schedule,
or use schedule(agent$get, event)
}
