% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_infection_functions.R
\name{t_msymp}
\alias{t_msymp}
\title{Mild infection duration}
\usage{
t_msymp(exp.im = 1/9)
}
\arguments{
\item{exp.im}{parameter of exponential distribution for time spent with mild symptoms}
}
\value{
numeric of time spent in mildly symptomatic (Im)
}
\description{
Draw from distribution of proposed length spent with minor symptoms. Used in `next.state`
}
