% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_testing_functions.R
\name{test_prob}
\alias{test_prob}
\title{Testing probability}
\usage{
test_prob(inf.status, income_bracket, age, res_type, res_inf, last_test)
}
\arguments{
\item{inf.status}{Infection status}

\item{income_bracket}{income}

\item{age}{age}

\item{res_type}{type of residence}

\item{res_inf}{residence infection status}

\item{last_test}{time since person was last tested}
}
\value{
sampling weight for whether or not tested
}
\description{
Estimate testing probability as function of agent characteristics and infection status
}
